refactor update profile view
