+++ backend/zane_api/views/serializers/__init__.py
@@ -1,4 +1,5 @@
 from .common import *
+from .auth import *
 from .environments import *
 from .deployments import *
 from .services import *
